==> 0_isw_data_collector.ipynb <==
# https://understandingwar.org/backgrounder/russian-offensive-campaign-assessment-march-9-2023

# BASE_URL = "https://understandingwar.org/backgrounder/russian-offensive-campaign-assessment"
# BASE_URL = 'https://isw.pub/RusCampaign'
# BASE_URL = 'https://isw.pub/UkrWar'

## Import necessary libraries and set-up variables

import requests
import time
from calendar import monthrange

OUTPUT_FOLDER = 'data/0_isw_data_collection/1_raw_isw_data'
BASE_URL = ""

def save_page(url, file_name):
    page = requests.get(url)
    
    url_name = url.split("/")[-1].replace("-","_")
    
    with open(f"{OUTPUT_FOLDER}/{file_name}___{url_name}.html",'wb+') as f:
        f.write(page.content)

mnths = ['Feb','Mar','Apr','May','June','July','August','Sept','Oct']

url = f"{BASE_URL}August26"
url

### feb25 - oct19 format

BASE_URL = 'https://isw.pub/RusCampaign'

YEAR = 2022
mon_ind = 2

for m in mnths:
    first_day = 1
    
    day_count = monthrange(YEAR, mon_ind)
    
    if m == 'Feb':
        first_day = 25
        
    for d in range(first_day,day_count[1]+1):
        if m == "Oct" and d>19:
            continue
        date = f"{m}{d}"
        
        file_name = f"{d}_{mon_ind}_{YEAR}"
        url = f"{BASE_URL}{date}"
        print(date)
        save_page(url,file_name)
        time.sleep(1)
    mon_ind +=1

### 2022 Oct,Now,Dec isw.pub format

BASE_URL = 'https://isw.pub/UkrWar'

YEAR = 2022
mon_ind = 10 

mnths = ['Oct','Now','Dec']

for m in mnths:
    first_day = 1
    
    day_count = monthrange(YEAR, mon_ind)
    #day_count[1]
    
    if m == 'Oct':
        first_day = 21
        
    for d in range(first_day,day_count[1]+1):
        if d <=9:
            sd = "0" + str(d)
        else:
            sd = str(d)
            
        date = f"{mon_ind}{sd}22"
        
        file_name = f"{d}_{mon_ind}_{YEAR}"
        url = f"{BASE_URL}{date}"
        
        print(date)
        save_page(url,file_name)
        time.sleep(1)
    mon_ind +=1

### 2023 january format



BASE_URL = "https://understandingwar.org/backgrounder/russian-offensive-campaign-assessment"

YEAR = 2023
mon_ind = 1

mnths = ["january"]



for m in mnths:
    first_day = 1
    
    day_count = monthrange(YEAR, mon_ind)
    
    for d in range(first_day,day_count[1]+1):
        if d <=9:
            sd = "0" + str(d)
        else:
            sd = str(d)
            
        file_name = f"{d}_{mon_ind}_{YEAR}"
        url = f"{BASE_URL}-{m}-{d}-{YEAR}"
        
        print(url)
        save_page(url,file_name)
        time.sleep(1)
        
    mon_ind +=1
