==> 0_isw_data_preparation.ipynb <==
## Import necessary libraries and read all isw report in html format

import pandas as pd 
import glob
from datetime import datetime 

from bs4 import BeautifulSoup

INPUT_DATA_FOLDER = "data/0_isw_data_collection/0_raw_isw_data"

#OUTPUT_FOlDER = "data/1_isw"
#OUTPUT_DATA_FILE = "all_days_isw_reports.csv"

files_by_days = glob.glob(f"{INPUT_DATA_FOLDER}/*.html")

len(files_by_days)

print(files_by_days[0])


### Generate Pandas DataFrame from 340 html pages using BeautifulSoup package

all_data = []

for file in files_by_days:
    
    d = {}
    file_name = file.split("/")[-1].split("___")
    date = datetime.strptime(file_name[0],"%d_%m_%Y")
    url = file_name[1].split(".")[0]
    
    with open(file,"r") as cfile:
        parsed_html = BeautifulSoup(cfile.read())
        
        title = parsed_html.head.find("title").text
        link = parsed_html.head.find("link",attrs={"rel":"canonical"},href=True)
        if parsed_html.head.find("link",attrs={"rel":"canonical"},href=True) is not None:
            link = parsed_html.head.find("link",attrs={"rel":"canonical"},href=True).attrs["href"]
        if parsed_html.body.find('h1', {'class': 'title', 'id': 'page-title'}) is not None:
            text_title = parsed_html.body.find('h1', {'class': 'title', 'id': 'page-title'}).text
        text_main = parsed_html.body.find('div',attrs={"class":"field field-name-body field-type-text-with-summary field-label-hidden"})
    #    print( parsed_html.body.find('h1', {'class': 'title', 'id': 'page-title'}).text)
        
        d = {
            'date': date,
            'short_url': url,
            'title': title,
            'text_title': text_title,
            'full_url': link,
            'main_html': text_main
        }
         
        all_data.append(d)

data = pd.DataFrame(all_data)

data['main_html'][1]

data = data.sort_values(by=['date'])

import requests

url = 'https://understandingwar.org/backgrounder/russian-offensive-campaign-assessment-march-9-2023'
response = requests.get(url)

html_code = response.content.decode('utf-8') #decode the content into a string

print(html_code)


data.head()

data.shape

data.isnull().count()

# extract raw dataset for further data preprocessing purposes

data.to_csv('data/0_isw_data_collection/0_raw_isw_parsed.csv')
