==> 2_isw_data_preparation.ipynb <==
## Import necessary libraries

import pandas as pd

data = pd.read_csv('data/0_isw_data_collection/1_isw_preprocessed.csv')

## Natural language preprocessing 

#!pip3 install --upgrade pip

#pip install num2words

#!pip install scikit-learn

# imports

from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import PorterStemmer
from nltk.stem import WordNetLemmatizer
from collections import Counter
from num2words import num2words


import nltk
import os
import string
import numpy as np
import copy 
import pickle
import re
import math

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.feature_extraction.text import TfidfTransformer


data.head()

## Prepare all  functions for NLP

def remove_one_letter_word(data):
    words = word_tokenize(str(data))
    
    new_text = ""
    for w in words:
        if (len(w) > 1):
            new_text = new_text + " " + w
            
    return new_text

def convert_lower_case(data):
    return np.char.lower(data)

nltk.download('stopwords')
stop_words = set(stopwords.words('english'))
print(stop_words)

def remove_stop_words(data):
    stop_words = set(stopwords.words('english'))
    stop_stop_words = {"no","not"}
    
    stop_words = stop_words - stop_stop_words
    
    words = word_tokenize(str(data))
    
    new_text = ""
    for w in words:
        if w not in stop_words and len(w) > 1:
            new_text = new_text + " " + w
            
    return new_text

def remove_punctuations(data):
    symbols = " !()-[]{};:'\,<>./?@#$%^&*_~ '"""  
    
    for i in range(len(symbols)):
        data = np.char.replace(data,symbols[i]," ")
        data = np.char.replace(data,"  "," ")
    
    data = np.char.replace(data,",", " ")
    
    return data
            

def stemming(data):
    stemmer = PorterStemmer()
    
    tokens = word_tokenize(str(data))
    new_text = ""
    for w in tokens:
        new_text = new_text + " " + stemmer.stem(w)
        
    return new_text

def lemmatizing(data):
    lemmatizer = WordNetLemmatizer()
    
    tokens = word_tokenize(str(data))
    new_text = ""
    for w in tokens:
        new_text = new_text + " " + lemmatizer.lemmatize(w)
               
    return new_text

def convert_numbers(data):
    
    tokens = word_tokenize(str(data))
    new_text = ""
    for w in tokens:
        if w.isdigit():
            if int(w) < 1000000000:
                w = num2words(w)
            else:
                w = " "
        new_text = new_text + " " + w
            
    new_text = np.char.replace(new_text,"-"," ")
        
    return new_text
                
                

def remove_url_from_string(data):
    words = word_tokenize(str(data))
    new_text = ""

    
    for w in words:
        w = re.sub(r'https?:\/\/.*[\r\n]*',"",str(w),flags=re.MULTILINE)
        w = re.sub(r'http?:\/\/.*[\r\n]*',"",str(w),flags=re.MULTILINE)
        
        new_text = new_text + " " + w
        
    return new_text

def preprocess(data, word_root_algo='lemm'):
    
    data = remove_one_letter_word(data)
    data = convert_lower_case(data)
    data = remove_stop_words(data)
    data = stemming(data)
    data = remove_punctuations(data)
    data = convert_numbers(data)
    data = remove_url_from_string(data)
    
    if word_root_algo =='lemm':
        data = lemmatizing(data)
    else: 
        data = stemming(data)
        
    
    data = remove_punctuations(data)
    data = remove_stop_words(data)
    
    return data

# apply all created functions with lemmatization

data['report_text_lemm'] = data['main_html_v6'].apply(lambda x: preprocess(x,"lemm"))

# apply all created functions with stemming 

data['report_text_stemm'] = data['main_html_v6'].apply(lambda x: preprocess(x,"stemm"))

data.head()

data['report_text_lemm'][0]

docs = data['report_text_lemm'].tolist()

data.to_csv('data/0_isw_data_collection/2_isw_preprocessed_before_vectosing.csv')

## CountVerctorizer algorithm

cv = CountVectorizer(max_df=0.98,min_df=2)
word_count_vector = cv.fit_transform(docs)

word_count_vector.shape

word_count_vector

with open("models/count_vectorizer_v1.pkl","wb") as handle:
    pickle.dump(cv, handle)

## TF-IDF algorithm

tfidf_transformer = TfidfTransformer(smooth_idf = True, use_idf = True)
tfidf_transformer.fit(word_count_vector)

with open("models/tfidf_transformer_v1.pkl","wb") as handle:
    pickle.dump(tfidf_transformer, handle)

df_idf = pd.DataFrame(tfidf_transformer.idf_, index = cv.get_feature_names(),columns=["idf_weights"])
df_idf.sort_values(by=['idf_weights'])

tf_idf_vector = tfidf_transformer.transform(word_count_vector)

tf_idf_vector

tf_idf_vector.shape

## Transformation with fitted models

tfidf = pickle.load(open("models/tfidf_transformer_v1.pkl","rb"))
cv = pickle.load(open("models/count_vectorizer_v1.pkl","rb"))

feature_names = cv.get_feature_names()

feature_names

def sort_coo(coo_matrix):
    
    tuples = zip(coo_matrix.col, coo_matrix.data)
    
    return sorted(tuples,key=lambda x: (x[1], x[0]), reverse=True)

def extract_topn_from_vector(feature_names, sorted_items, topn=10):
    
    sorted_items = sorted_items[:topn]
    
    score_vals = []
    feature_vals = []
    
    for idx, score in sorted_items:
        
        score_vals.append(round(score,3))
        feature_vals.append(feature_names[idx])
        
    results = {}
    
    for idx in range(len(feature_vals)):
        results[feature_vals[idx]] = score_vals[idx]
    
    return results


def convert_doc_to_vector(doc):
    feature_names = cv.get_feature_names()
    top_n = 10
    tf_idf_vector = tfidf.transform(cv.transform([doc]))
    
    sorted_items = sort_coo(tf_idf_vector.tocoo())
    
    keywords = extract_topn_from_vector(feature_names, sorted_items, top_n)
    
    return keywords

data['keywords'] = data['report_text_stemm'].apply(lambda x: convert_doc_to_vector(x))

data['keywords'][0]

data_vectorised = tf_idf_vector.toarray()

vectors_df = pd.DataFrame(data_vectorised)

vectors_df.to_csv('data/0_isw_data_collection/3_isw_vectorised.csv')

vectors_df.head()

vectors_df.shape

vectors_df['date'] = data['date']

vectors_df.head(10)

# extract dataset for further purposes

vectors_df.to_csv('data/0_isw_data_collection/3_isw_vectorised_with_date.csv')
