==> 3_feature_engineering.ipynb <==
## Import necessary libraries

import pandas as pd

alarms = pd.read_csv("data/0_raw_other_data/alarms.csv",sep=";")

alarms.head()

alarms['start_time'] = pd.to_datetime(alarms['start'])
alarms['end_time'] = pd.to_datetime(alarms['end'])
alarms["day"] = alarms['start_time'].dt.day

# the question is that could be the alarm longer than 24 hours?

alarms.head(40)

## Generating first feauture: A number of regions where the alarm is going on

alarms['feature_number_of_region'] = 0
for i in range(0,alarms.shape[0]):
    count=0
    for j in range(0,alarms.shape[0]):
        if (alarms['start_time'][j] >= alarms['start_time'][i]) and (alarms['end_time'][j] < alarms['end_time'][i]):
            count = count+1
        if alarms['day'][j] > alarms['day'][j+1]:
            break
    alarms['feature_number_of_region'][i] = count
    count=0
        

alarms.head(20)

#generating first feauture: Number of alarms for this region during the last 24 hours;
alarms['within_24_hours']=0
for i in range(0,alarms.shape[0]): 
    within_24_hours = alarms['start_time'][i] - pd.Timedelta(hours=24)
    alarms['within_24_hours'][i] = within_24_hours


alarms.head()

alarms = alarms.sort_values(by=['region_title'])

alarms.to_csv('alarms_sorted.csv')

alarms.head()

## Generating second feature: Number of alarms for this region during the last 24 hours

alarms['feature_number_of_alarms_within_24_hours'] = 0
for i in range(0,alarms.shape[0]):
    count=0
    for j in range(0,alarms.shape[0]):
        if ((alarms['within_24_hours'][i]) < (alarms['start_time'][j]) < (alarms['start_time'][i])):
            count = count+1
        if(alarms['region_title'][j] != alarms['region_title'][j+1]):
            break
    alarms['feature_number_of_alarms_within_24_hours'][i] = count
    count=0

alarms.head(50)

alarms['feature_number_of_alarms_within_24_hours'].value_counts()

regions = alarms['region_title'].unique()

print(regions)

#data_1 = alarms[alarms['region_title'] == regions[0]]

#data_1.head()

alarms['start_time'][1] - alarms['start_time'][2] < pd.Timedelta(hours=24)

# export dataset for further purposes

alarms.to_csv('data/alarms_with_features.csv')
