==> 5_evaluate_models.ipynb <==
## Import necessary libraries and models

import pickle
import numpy as np
import pandas as pd

knn = pickle.load(open("models/training_models/4_knn_1.1.pkl","rb"))
rf = pickle.load(open("models/training_models/4_rf_1.1.pkl","rb"))

## Extract feature importance

X = pd.read_csv('data/weather_alarms_tfidf_features_clear.csv')
X = X.drop(columns=['Unnamed: 0.1', 'Unnamed: 0'], axis=1)

feature_importances = rf.feature_importances_

# Sort the feature importances in descending order
sorted_idx = np.argsort(feature_importances)[::-1]

for i in range(20):
    print(feature_importances[sorted_idx[i]], X.columns[sorted_idx[i]])

# 0.07056355847478037 day_datetimeEpoch
# 0.06315651241432549 hour_dew
# 0.06192005341819136 hour_pressure
# 0.06135555793062799 hour_humidity
# 0.06062265469245814 hour_windspeed
# 0.059016038625131884 hour_precipprob
# 0.05779101646456585 hour_temp
# 0.057750866882785336 hour_winddir
# 0.04892916520458762 hour_visibility
# 0.04758592836946624 hour_severerisk
# 0.03355601202311686 day_solarenergy
# 0.03352069441012746 day_precip
