==> 5_predict_alarms.ipynb <==
## Import libraries

# import libraries

import datetime
import pickle
import nltk
import re
import scipy
import pytz
import pandas as pd
from scipy import sparse
from bs4 import BeautifulSoup

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
from utils import get_weather
from utils import text_processing

## Get and preprocess ISW files

# get article from yesterday

today = datetime.date.today()
yesterday = today - datetime.timedelta(days=1)

yesterday_day = yesterday.day
yesterday_month = yesterday.month
yesterday_year = yesterday.year

file = text_processing.get_article_from_yesterday(yesterday_day,yesterday_month,yesterday_year)

data = text_processing.read_html(file)

def preprocess_all_text(data):
    pattern = "\[(\d+)\]"
    data['main_html_v1'] = data['main_html'].apply(lambda x: re.sub(pattern,"",str(x)))
    data['main_html_v2'] = data['main_html_v1'].apply(lambda x: re.sub(r'http(\S+.*\s)',"",x))
    data['main_html_v3'] = data['main_html_v2'].apply(lambda x: re.sub(r'2022|2023|©2022|©2023|\xa0|\n',"",x))
    data['main_html_v4'] = data['main_html_v3'].apply(lambda x: BeautifulSoup(x).text)
    data['main_html_v5'] = data['main_html_v4'].apply(lambda x: text_processing.remove_names_and_dates(x))
    
    return data

data_preprocessed = preprocess_all_text(data)

data_preprocessed = data_preprocessed.drop(['main_html_v1','main_html_v2','main_html_v3','main_html_v4'],axis=1)

# Uncommit for first time download

# nltk.download('stopwords')
# nltk.download('punkt')
# nltk.download('wordnet')

# lemmatization and stemming algorithms to get the unique words

data_preprocessed['report_text_lemm'] = data_preprocessed['main_html_v5'].apply(lambda x: text_processing.preprocess(x,"lemm"))
data_preprocessed['report_text_stemm'] = data_preprocessed['main_html_v5'].apply(lambda x: text_processing.preprocess(x,"stemm"))

data_preprocessed.head()

docs = data_preprocessed['report_text_lemm'].tolist()

## ISW vectorize

# apply vectorizing and tf-idf algorithm

cv = CountVectorizer()
word_count_vector = cv.fit_transform(docs)

word_count_vector.shape

tfidf_transformer = TfidfTransformer(smooth_idf = True, use_idf = True)
tfidf_transformer.fit(word_count_vector)

tf_idf_vector = tfidf_transformer.transform(word_count_vector)

tf_idf_vector

# extract the most valuable words in article

feature_names = cv.get_feature_names_out()
tf_idf_vector
data_preprocessed['keywords'] = data_preprocessed['report_text_stemm'].apply(lambda x: text_processing.convert_doc_to_vector(x,feature_names,tf_idf_vector))

data_preprocessed['keywords'][0]

data_preprocessed.head()

## Part of script: Final preprocessing and merging

data_preprocessed["date_datetime"] = pd.to_datetime(data_preprocessed["date"])
data_preprocessed['date_tomorrow_datetime'] = data_preprocessed['date_datetime'].apply(lambda x: x+datetime.timedelta(days=1))
data_preprocessed = data_preprocessed.rename(columns = {"date_datetime":"report_date"})

data_preprocessed.head()

data_vectorised = tf_idf_vector.toarray()
vectors_df = pd.DataFrame(data_vectorised)
vectors_df['date'] = pd.to_datetime(today)


vectors_df.head()

df_isw_short = data_preprocessed[['date','report_text_lemm','keywords','date_tomorrow_datetime']]

df_isw_short.head()

## Get weather and make predictions for all regions

# load necessary models

tfidf = pickle.load(open("models/tfidf_transformer_v1.pkl","rb"))
cv = pickle.load(open("models/count_vectorizer_v1.pkl","rb"))
model = pickle.load(open("models/training_models/4_rf_3.1f.pkl","rb"))

cities = ['Kyiv', 'Kharkiv', 'Lviv', 'Odesa', 'Donetsk', 'Luhansk', 'Kherson', 'Simferopol', 'Mykolaiv', 'Lutsk', 'Chernihiv', 'Dnipro', 'Vinnytsia', 'Zhytomyr', 'Kropyvnytskyi', 'Poltava', 'Sumy', 'Rivne', 'Khmelnytskyi', 'Chernivtsi', 'Ternopil', 'Cherkasy', 'Uzhgorod', 'Zaporozhye', 'Ivano-Frankivsk']

date = datetime.datetime.now(pytz.timezone('Europe/Kyiv'))
result = {}

# for loop generate weather, merged it with isw vector and make a prediction for all regions in list above

for city in cities:

    df_weather_complete = get_weather.get_weather_for_12_hours(city,date)

    # merge
    df_weather_complete['key']=1
    df_isw_short['key']=1
    df_all = df_weather_complete.merge(df_isw_short, how = 'left', left_on = 'key', right_on = 'key')

    # drop
    to_drop=['key','date','date_tomorrow_datetime','keywords','report_text_lemm']
    if 'sunrise' in df_all.columns:
        exceptions = ['sunset','sunrise']
        to_drop.extend(exceptions)
    df_weather_matrix = df_all.drop(to_drop, axis = 1)

    # final dataset
    df_weather_matrix['Unnamed: 0'] = 0
    df_weather_matrix= df_weather_matrix[['Unnamed: 0', 'day_datetimeEpoch', 'day_tempmax', 'day_tempmin', 'day_temp', 'day_dew', 'day_humidity', 'day_precip', 'day_precipcover', 'day_solarradiation', 'day_solarenergy', 'day_uvindex', 'day_moonphase', 'hour_datetimeEpoch', 'hour_temp', 'hour_humidity', 'hour_dew', 'hour_precipprob', 'hour_snow', 'hour_snowdepth', 'hour_windgust', 'hour_windspeed', 'hour_winddir', 'hour_pressure', 'hour_visibility', 'hour_cloudcover', 'hour_severerisk', 'region_id']]

    cv_vector_model = cv.transform(df_all['report_text_lemm'].values.astype('U'))
    tf_idf_model = tfidf.transform(cv_vector_model)

    # merge final dataset with tfidf vector
    df_weather_matrix_csr = scipy.sparse.csr_matrix(df_weather_matrix)
    df_all_data_csr = scipy.sparse.hstack((df_weather_matrix_csr, tf_idf_model), format='csr')

    #predict
    predicted = model.predict(df_all_data_csr)
    hours = []
    
    # store results in dictionary, after extract 
    for i in range(12):
        hour = date + datetime.timedelta(hours=i)
        hour_rounded = hour.replace(minute=0, second=0, microsecond=0)
        hours.append(hour_rounded.strftime('%Y-%m-%d %H:%M'))

    result[city] = dict(zip(hours, predicted))

#Show result
result

##     Save results

result = pd.DataFrame(result)
VERSION = "3"
result.to_csv(f'data/results/results_{VERSION}.txt', sep='\t', index=False)
